RootSignature 中分别存储 CBV_SRV_UAV 和 SAMPLER 的数量 CommandList 中的 DynamicDescriptorHeap 调整为 2 个, CBV_SRV_UAV 和 SAMPLER DescriptorAllocatorPage 中的编译警告

+++ LearnDirectX12/Component/dx12lib/DescriptorAllocatorPage.cpp
@@ -72,7 +72,7 @@ DescriptorAllocation DescriptorAllocatorPage::allocate(uint32 numDescriptor) {
 	
 	_freeListByOffset.erase(offset);
 	_freeListBySize.erase(iter);
-	uint32 newSize = size - numDescriptor;
+	uint32 newSize = static_cast<UINT>(size) - numDescriptor;
 	uint32 newOffset = offset + numDescriptor;
 	_numFreeHandle -= numDescriptor;
 	if (newSize > 0)
